pyproject.toml
[tool.poetry]
name = "moviebuff"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
imdb-cli-tool = "^1.0.1"
bs4 = "^0.0.1"
imageio = "^2.9.0"
scikit-image = "^0.18.1"
flask = "^1.1.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
